Migration to .Net 8 and small fixes
Fixed the calls to Windows Audio API that passed the pointer to In and Our audio buffer as Int32 instead of IntPtr. This was causing a blocking error.
Fixed references to Windows.Forms. instead of System,Windows.Forms. or nothing, in the custom controls.
A fresh .Net 8 empty project has been generated; then the files from the previous project have been copied in the project's folder. Fixes to the new project have been done.
Program's info contained in the former AssemblyInfo.vb file have been moved to the "packet" properties of the program.
The structure of the folders of the Solution has been changed from the now "standard" .Net to the old fashioned  "pre .Net Core" model, to better match the previous project.